Refactoring Dagger implementation
The way Dagger was implemented, it wasn't easy to do constructor injector neither having Application as a dependency to be injected.

So, I decided to refactor, creating a new ApplicationComponent, which built on CustomApplication (before any Activity being attached).

Another thing that I did was refactoring the ViewModelFactory: this new one is generic and doesn't need boiler plate code when you create a new fragment. The only thing you'll need to do, when adding a new ViewModel to the project, is adding it to the ViewModelModule and you're good to go.